Refactoring and Modular code
